requirements.txt
openai
colorama
mysql-connector-python
chevron
pyyaml